Move ThirdPartyDataRetrievalMetadata related models
