+++ src/server/RobinTTY.PersonalFinanceDashboard.API/appsettings.Development.json
@@ -8,7 +8,7 @@
       "Default": "Information",
       "Override": {
         "Microsoft.AspNetCore": "Warning",
-        "Microsoft.EntityFrameworkCore": "Warning"
+        "Microsoft.EntityFrameworkCore": "Information"
       }
     },
     "WriteTo": [
