+++ src/server/RobinTTY.PersonalFinanceDashboard.Core/Models/ThirdPartyDataType.cs
