+++ src/server/RobinTTY.PersonalFinanceDashboard.Infrastructure/Repositories/ThirdPartyDataRetrievalMetadataRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using RobinTTY.PersonalFinanceDashboard.Core.Models;
 using RobinTTY.PersonalFinanceDashboard.Infrastructure.Entities;
 using RobinTTY.PersonalFinanceDashboard.ThirdPartyDataProviders.Models;
